backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
redis>=5.0.8
celery>=5.4.0
djangorestframework>=3.15.2
aiogram>=3.10.0
gunicorn>=22.0.0
whitenoise>=6.8.2
